@@ -14,41 +14,58 @@
     "author": "Roman Dvornov <rdvornov@gmail.com> (https://github.com/lahmatiy)",
     "license": "MIT",
     "repository": "discoveryjs/json-ext",
-    "main": "./src/index",
+    "type": "module",
+    "sideEffects": false,
     "browser": {
-        "./src/stringify-info.js": "./src/stringify-info.js",
-        "./src/parse-chunked.js": "./src/parse-chunked.js",
-        "./src/stringify-stream.js": "./src/stringify-stream-browser.js",
-        "./src/text-decoder.js": "./src/text-decoder-browser.js",
-        "./src/version.js": "./dist/version.js"
+        ".": "./dist/browser.cjs",
+        "./src/stringify-stream": "./src/stringify-stream-browser",
+        "./src/text-decoder": "./src/text-decoder-browser"
     },
+    "exports": {
+        "./package.json": {
+            "default": "./package.json"
+        },
+        "./dist/browser": {
+            "default": "./dist/browser.cjs"
+        },
+        "./dist/browser.mjs": {
+            "default": "./dist/browser.mjs"
+        },
+        "./dist/browser.global": {
+            "default": "./dist/browser.global.js"
+        },
+        ".": {
+            "require": "./dist/index.cjs",
+            "node": "./dist/index.cjs",
+            "import": "./dist/index.mjs",
+            "default": "./dist/index.mjs"
+        }
+    },
+    "main": "./src/index",
     "types": "./index.d.ts",
     "scripts": {
         "test": "mocha --reporter progress",
         "lint": "eslint src test",
         "lint-and-test": "npm run lint && npm test",
-        "build": "rollup --config",
-        "test:all": "npm run test:src && npm run test:dist",
-        "test:src": "npm test",
-        "test:dist": "cross-env MODE=dist npm test && cross-env MODE=dist-min npm test",
-        "build-and-test": "npm run build && npm run test:dist",
-        "coverage": "c8 --reporter=lcovonly npm test",
+        "build": "node esbuild.js",
+        "test:all": "npm run test && npm run test:dist",
+        "test:dist": "npm run test:node && npm run test:browser",
+        "test:node": "cross-env MODE=node npm test",
+        "test:browser": "cross-env MODE=browser npm test",
+        "build-and-test": "npm run build && npm run test:all",
+        "coverage": "c8 --reporter=text --reporter=lcovonly npm test",
         "prepublishOnly": "npm run lint && npm test && npm run build-and-test"
     },
     "devDependencies": {
-        "@rollup/plugin-commonjs": "^15.1.0",
-        "@rollup/plugin-json": "^4.1.0",
-        "@rollup/plugin-node-resolve": "^9.0.0",
-        "c8": "^7.10.0",
-        "chalk": "^4.1.0",
+        "esbuild": "^0.20.1",
+        "c8": "^9.1.0",
+        "chalk": "^5.3.0",
         "cross-env": "^7.0.3",
-        "eslint": "^8.10.0",
-        "mocha": "^9.2.1",
-        "rollup": "^2.28.2",
-        "rollup-plugin-terser": "^7.0.2"
+        "eslint": "^8.57.0",
+        "mocha": "^10.3.0"
     },
     "engines": {
-        "node": ">=10.0.0"
+        "node": ">=16.0.0"
     },
     "files": [
         "dist",
