@@ -1,3 +1,3 @@
-module.exports = () => {
+export default () => {
     throw new Error('Method is not supported');
 };
