@@ -1,6 +1,4 @@
-module.exports = {
-    version: require('./version'),
-    stringifyInfo: require('./stringify-info'),
-    stringifyStream: require('./stringify-stream'),
-    parseChunked: require('./parse-chunked')
-};
+export {default as version} from './version.js';
+export {default as stringifyInfo} from './stringify-info.js';
+export {default as stringifyStream} from './stringify-stream.js';
+export {default as parseChunked} from './parse-chunked.js';
