@@ -127,18 +127,21 @@ function normalizeSpace(space) {
     return false;
 }
 
-module.exports = {
+const type = {
+    PRIMITIVE: PrimitiveType,
+    PROMISE: PromiseType,
+    ARRAY: ArrayType,
+    OBJECT: ObjectType,
+    STRING_STREAM: ReadableStringType,
+    OBJECT_STREAM: ReadableObjectType
+};
+
+export {
     escapableCharCodeSubstitution,
     isLeadingSurrogate,
     isTrailingSurrogate,
-    type: {
-        PRIMITIVE: PrimitiveType,
-        PROMISE: PromiseType,
-        ARRAY: ArrayType,
-        OBJECT: ObjectType,
-        STRING_STREAM: ReadableStringType,
-        OBJECT_STREAM: ReadableObjectType
-    },
+
+    type,
 
     isReadableStream,
     replaceValue,
